+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Fixed
+
+- Fixed RLS policies with long table names causing duplicate policies. ([#2](https://github.com/cfeenstra67/sqlauthz/pull/2), thanks @pnispel)
+
 ## [1.0.2] - 2024-07-06
 
 ### Added